+++ App.js
@@ -1,7 +1,23 @@
 import LoginScreen from './src/views/LoginScreen';
+import AfterLogin from './src/views/AfterLogin';
+import {NavigationContainer} from '@react-navigation/native';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+const Stack = createNativeStackNavigator();
 
 export default function App() {
   return (
-    <LoginScreen/>
+    <NavigationContainer>
+      <Stack.Navigator 
+        screenOptions={{
+          headerShown: false
+          }}
+      >
+        <Stack.Screen
+          name="Login"
+          component={LoginScreen}
+        />
+        <Stack.Screen name="AfterLogin" component={AfterLogin} />
+      </Stack.Navigator>
+    </NavigationContainer>
   );
 }
